add memory pool capacity check and transaction expiration check
